commenting on files (3 missing) + admin only area

+++ deleteCategory.php
@@ -1,20 +1,31 @@
 <?php
 session_start();
+/*The requires are used to generate the templates as well as connect to the database
+These are not stored in the public directory as they user should not be able to directly access these*/
 $title = 'NN - Delete Category';
 require '../head.php';
 require '../nav.php'; 
-require '../sideNavBar.php';
 require '../databaseJoin.php';
 
-if(isset($_GET['name'])) {
-    $removeCategory = $pdo->prepare('DELETE FROM category WHERE name= :name');
-    $removeCategory ->execute($_GET);
+//The if statement below checks if an admin is logged in, the code for the page will only run if an admin is logged in
+if(isset($_SESSION['adminloggedin'])) {
 
-echo '<p>Category Deleted</p>';
+	/*the side bar require is not with the rest of the templates as access to this could 
+    allow non-admins to get access to admin only areas */
+	require '../sideNavBar.php';
+
+    /*When the category delete is clicked the name is used to ensure the right article gets deleted*/
+    if(isset($_GET['name'])) {
+        $removeCategory = $pdo->prepare('DELETE FROM category WHERE name= :name');
+        $removeCategory ->execute($_GET);
+
+    echo '<p>Category Deleted</p>';
+    }
 }
-?>
-
-<!-- contains the page footer and the closing html -->
-<?php
+//if an admin is not logged in then error messages are echo'd asking for a admin log in to continue
+else {
+    echo '<h3>You need to be logged in as a Admin to access this page</h3>';
+    echo '<h3>Please Log in as a Admin to try again</h3>';
+}
 require '../foot.php';
 ?>
